requirements.txt
flask==2.0.1
werkzeug==2.0.3
huggingface_hub==0.16.4
requests==2.28.2
nltk==3.7
numpy==1.24.3
scikit-learn==1.2.2
python-dotenv==0.19.2 